Fix for #483
Changed core._request_loader so that api tokens generated with version 3.x are correctly consumed
