+++ tests/test_common.py
@@ -21,9 +21,10 @@ from flask_security.utils import hash_data
 
 from tests.test_utils import (
     authenticate,
-    json_authenticate,
+    get_auth_token_version_3x,
     get_num_queries,
     hash_password,
+    json_authenticate,
     logout,
     populate_data,
     verify_token,
@@ -751,20 +752,14 @@ def test_change_uniquifier(app, client_nc):
     verify_token(client_nc, token)
 
 
-def test_verifying_token_from_version_3x(app, client_nc):
+def test_verifying_token_from_version_3x(in_app_context):
     """
     Check token generated with flask security 3.x, which has different form
     than token from version 4.0.0, can be verified
     """
 
-    def get_auth_token_version_3x(app, user):
-        """
-        Copy of algorithm that generated user token in version 3.x
-        """
-        data = [str(user.id), hash_data(user.password)]
-        if hasattr(user, "fs_uniquifier"):
-            data.append(user.fs_uniquifier)
-        return app.security.remember_token_serializer.dumps(data)
+    app = in_app_context
+    populate_data(app)
 
     with app.test_request_context("/"):
         user = app.security.datastore.find_user(email="matt@lp.com")
@@ -937,3 +932,25 @@ def test_no_get_auth_token(app, client):
         headers={"Content-Type": "application/json"},
     )
     assert "authentication_token" not in response.json["response"]["user"]
+
+
+def test_auth_token_decorator(in_app_context):
+    """
+    Test accessing endpoint decorated with auth_token_required
+    when using token generated by flask security 3.x algorithm
+    """
+
+    app = in_app_context
+    populate_data(app)
+    client_nc = app.test_client(use_cookies=False)
+
+    with app.test_request_context("/"):
+
+        user = app.security.datastore.find_user(email="matt@lp.com")
+        token = get_auth_token_version_3x(app, user)
+
+        response = client_nc.get(
+            "/token",
+            headers={"Content-Type": "application/json", "Authentication-Token": token},
+        )
+        assert response.status_code == 200
